Don't upload symbols to DO anymore
