+++ .github/workflows/danger.yml
@@ -28,10 +28,10 @@ jobs:
         node-version: '20'
         cache: pnpm
         cache-dependency-path: script/danger/pnpm-lock.yaml
-    - name: steps::danger::install_deps
+    - name: danger::install_deps
       run: pnpm install --dir script/danger
       shell: bash -euxo pipefail {0}
-    - name: steps::danger::run
+    - name: danger::run
       run: pnpm run --dir script/danger danger ci
       shell: bash -euxo pipefail {0}
       env:
