+++ tooling/xtask/src/tasks/workflows.rs
@@ -3,12 +3,13 @@ use clap::Parser;
 use std::fs;
 use std::path::Path;
 
+mod danger;
+mod nix;
+mod run_bundling;
+
 mod runners;
 mod steps;
 mod vars;
-mod workflows;
-
-use workflows::*;
 
 #[derive(Parser)]
 pub struct GenerateWorkflowArgs {}
@@ -17,9 +18,9 @@ pub fn run_workflows(_: GenerateWorkflowArgs) -> Result<()> {
     let dir = Path::new(".github/workflows");
 
     let workflows = vec![
-        ("danger.yml", danger()),
-        ("nix.yml", nix()),
-        ("run_bundling.yml", run_bundling()),
+        ("danger.yml", danger::danger()),
+        ("nix.yml", nix::nix()),
+        ("run_bundling.yml", run_bundling::run_bundling()),
     ];
     fs::create_dir_all(dir)
         .with_context(|| format!("Failed to create directory: {}", dir.display()))?;
