+++ tooling/xtask/src/tasks/workflows/steps.rs
@@ -57,113 +57,28 @@ pub fn upload_artifact(name: &str, path: &str) -> Step<Use> {
 
 // todo! we used to do 300Mb on macOS and 100Mb on Windows
 pub fn clean_target_dir() -> Step<Run> {
-    named::run("script/clear-target-dir-if-larger-than 100")
+    named::bash("script/clear-target-dir-if-larger-than 100")
 }
 
 pub fn script(name: &str) -> Step<Run> {
-    Step::new(name).run(name).shell(BASH_SHELL)
-}
-
-// todo! can probably combine with script with `if name.ends_with(".ps1")`
-pub fn script_windows(name: &str) -> Step<Run> {
-    Step::new(name).run(name).shell(POWERSHELL_SHELL)
-}
-
-pub mod danger {
-    use super::*;
-
-    pub fn setup_node() -> Step<Use> {
-        super::setup_node()
-            .add_with(("cache", "pnpm"))
-            .add_with(("cache-dependency-path", "script/danger/pnpm-lock.yaml"))
-    }
-
-    pub fn install_deps() -> Step<Run> {
-        named::run("pnpm install --dir script/danger")
-    }
-
-    pub fn run() -> Step<Run> {
-        named::run("pnpm run --dir script/danger danger ci")
-            // This GitHub token is not used, but the value needs to be here to prevent
-            // Danger from throwing an error.
-            .add_env(("GITHUB_TOKEN", "not_a_real_token"))
-            // All requests are instead proxied through an instance of
-            // https://github.com/maxdeviant/danger-proxy that allows Danger to securely
-            // authenticate with GitHub while still being able to run on PRs from forks.
-            .add_env((
-                "DANGER_GITHUB_API_BASE_URL",
-                "https://danger-proxy.fly.dev/github",
-            ))
+    if name.ends_with(".ps1") {
+        Step::new(name).run(name).shell(POWERSHELL_SHELL)
+    } else {
+        Step::new(name).run(name).shell(BASH_SHELL)
     }
 }
 
-pub mod nix {
-    use indoc::indoc;
-
-    use crate::tasks::workflows::vars;
-
-    use super::*;
-
-    // on our macs we manually install nix. for some reason the cachix action is running
-    // under a non-login /bin/bash shell which doesn't source the proper script to add the
-    // nix profile to PATH, so we manually add them here
-    pub fn set_path() -> Step<Run> {
-        named::run(indoc! {r#"
-            echo "/nix/var/nix/profiles/default/bin" >> "$GITHUB_PATH"
-            echo "/Users/administrator/.nix-profile/bin" >> "$GITHUB_PATH"
-        "#})
-    }
-
-    pub fn install_nix() -> Step<Use> {
-        named::uses(
-            "cachix",
-            "install-nix-action",
-            "02a151ada4993995686f9ed4f1be7cfbb229e56f", // v31
-        )
-        .add_with(("github_access_token", vars::GITHUB_TOKEN))
-    }
-
-    pub fn cachix_action(cachix_filter: &str) -> Step<Use> {
-        named::uses(
-            "cachix",
-            "cachix-action",
-            "0fc020193b5a1fa3ac4575aa3a7d3aa6a35435ad", // v16
-        )
-        .add_with(("name", "zed"))
-        .add_with(("authToken", vars::CACHIX_AUTH_TOKEN))
-        .add_with(("pushFilter", cachix_filter))
-        .add_with(("cachixArgs", "-v"))
-    }
-
-    pub fn build(flake_output: &str) -> Step<Run> {
-        named::run(&format!(
-            "nix build .#{} -L --accept-flake-config",
-            flake_output
-        ))
-    }
-
-    pub fn limit_store() -> Step<Run> {
-        named::run(indoc! {r#"
-            if [ "$(du -sm /nix/store | cut -f1)" -gt 50000 ]; then
-                nix-collect-garbage -d || true
-            fi"#
-        })
-    }
-}
-
-// (janky) helpers to generate steps with a name that corresponds
+// (janky) helper to generate steps with a name that corresponds
 // to the name of the calling function.
-mod named {
-    use gh_workflow::*;
+pub(crate) mod named {
+    use super::*;
 
-    pub(super) fn uses(owner: &str, repo: &str, ref_: &str) -> Step<Use> {
+    pub(crate) fn uses(owner: &str, repo: &str, ref_: &str) -> Step<Use> {
         Step::new(function_name(1)).uses(owner, repo, ref_)
     }
 
-    pub(super) fn run(script: &str) -> Step<Run> {
-        Step::new(function_name(1))
-            .run(script)
-            .shell(super::BASH_SHELL)
+    pub(crate) fn bash(script: &str) -> Step<Run> {
+        Step::new(function_name(1)).run(script).shell(BASH_SHELL)
     }
 
     fn function_name(i: usize) -> String {
@@ -182,7 +97,8 @@ mod named {
             false
         });
         name.split("::")
-            .skip_while(|s| s != &"steps")
+            .skip_while(|s| s != &"workflows")
+            .skip(1)
             .collect::<Vec<_>>()
             .join("::")
     }
